Refactor LLM Task helper
