+++ src/zrb/task/llm/typing.py
@@ -1,0 +1,3 @@
+from typing import Any
+
+ListOfDict = list[dict[str, Any]]
